requirements.txt
# ETL Learning Project Dependencies

# Apache Spark
pyspark==3.5.0

# Kafka
kafka-python==2.0.2

# Database Connectors
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy==2.0.25

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Monitoring & Logging
colorlog==6.8.0

# Utilities
faker==22.0.0  # For realistic data generation
